Notebooks/requirements.txt
# .ipynb
nbformat
ipykernel
nb-clean

# basic python
mypy
python-dotenv
scikit-learn
scipy
numpy
pandas
faker
# displaying
plotly
seaborn
matplotlib
tqdm

# BERTopic / Transformers
bertopic
keybert
transformers
huggingface_hub
datasets

# pre-procecssing
beautifulsoup4
emoji
lxml

# google
google-genai
google-generativeai
